Use an improved comparison function for "debug" tests
The current method does a simple "diff", which means
that differences in line order appear as failures.
Since output is being redirected, there is no
guarantee of ordering. We have been dealing with
ordering-related errors for some time now, resolving
them by simply retrying until the order comes out
"right".

Switch to a custom comparison program that checks
to see that all lines appear in the new output
when compared to the baseline, and that they
appear the expected number of times. Output
any outliers and return an error status.

Signed-off-by: Ralph Castain <rhc@pmix.org>