Update the xversion tests
We haven't supported v2.x in ages, and it isn't included in
the list of supported_versions - so remove it from the script.
Since only old unsupported PMIx versions require HWLOC v1.x,
let's not include that version in our tests so this becomes
more focused on just xversion support.

Signed-off-by: Ralph Castain <rhc@pmix.org>